Non auto detect board

+++ lib/boards/soc/BCM2712/gpio/gpio_BCM2712.c
@@ -4,7 +4,7 @@
 void bcm2712_init_pin(uint8_t pin, gpio_mode_t mode, gpio_pud_t pud) { /* ... */ }
 void bcm2712_set(uint8_t pin) { /* ... */ }
 void bcm2712_clear(uint8_t pin) { /* ... */ }
-uint8_t bcm2712_read(uint8_t pin) { /* ... */ return 0;}
+uint32_t bcm2712_read(uint8_t pin) { /* ... */ return 0;}
 
 const gpio_ops_t bcm2712_gpio_ops = {
     .init_pin = bcm2712_init_pin,
